% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{hai_clean}
\alias{hai_clean}
\title{HAI cases (Germany, PPS 2011)}
\format{
A tibble with 5 rows and 3 variables:
\describe{
\item{infection}{Character, infection type: HAP, SSI, BSI, UTI, CDI}
\item{cases}{Integer, number of cases}
\item{country}{Character, always "Germany"}
}
}
\source{
Zacher et al. (2019), via the BHAI package.
}
\usage{
hai_clean
}
\description{
Tidy counts of healthcare-associated infections (HAIs) in Germany.
Created from \code{BHAI::num_hai_patients}.
}
\examples{
data(hai_clean)
head(hai_clean)
}
\keyword{datasets}
