% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Shiny app}
\usage{
run_app()
}
\value{
Invisible NULL; runs the app for its side-effect.
}
\description{
This launches the interactive Shiny application included in the haiDash package.
}
